Change factorial type from long to int

+++ Factorial.java
@@ -1,12 +1,12 @@
 public class Factorial {
     public static void main(String[] args) {
         int number = 5;
-        long factorial = calculateFactorial(number);
+        int factorial = calculateFactorial(number);
         System.out.println("Factorial of " + number + " is: " + factorial);
     }
 
-    public static long calculateFactorial(int n) {
-        long result = 1;
+    public static int calculateFactorial(int n) {
+        int result = 1;
         for (int i = 1; i <= n; i++) {
             result *= i;
         }
